Change mystery to two pages
